pkg/operator: move sync file to sync.go
